# swagger.yaml
swagger: "2.0"
info:
  description: "This is the Team 1 group project for Boston University's MET CS 673 class in Summer 1 2022 (May - June).  The application is modeled after personal finance applications like Mint and Truebill and allows users to input bank account and transaction data.  This data can then be accessed, modified, or voided at the user's discretion and it allows them to see their spending habits at a high level. \n\n Team members: Jean Dorancy, Timothy Flucker, Xiaobing Hou, Weijie Liang"
  version: "1.0.0"
  title: "Trackr - Personal Transaction Management"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "tflucker@bu.edu"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "trackr.swagger.io"
basePath: "/v1"
...
schemes:
  - "https"
  - "http"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /register:
    post:
      tags: 
        - "User Management"
      summary: "Creates a new user with the provided data. Returns a JWT Token value."
      description: "Unsecured API Endpoint that allows for new users to be created in the system. Upon valid inputs, a new user record will be created and a new JWT Token will be returned to the user for future API calls."
      operationId: "register"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TrackrUserDTO"
      responses:
        "200": 
          description: "success response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "SUCCESS"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Successfully created new JWT token!"
              additionalData:
                type: "string"
                example: "eyJhbGciOiJIUzI1NiJ9.eyJuYW1lIjoiSm9lIENvZGVyIn0.5dlp7GmziL2QS06sZgK4mtaqv0_xX4oFUuTDh1zHK4U"
  /login:
    post:
      tags: 
        - "User Management"
      summary: "Validate if that user exists using the provided login credentials and return a JWT Token value."
      description: "Unsecured API Endpoint that allows users to enter the system (front-end) or generate a new JWT Token if they forgot their previous token or if it expired."
      operationId: "login"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TokenRetrievalDTO"
      responses:
        "200": 
          description: "success response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "SUCCESS"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Successfully created new JWT token!"
              additionalData:
                type: "string"
                example: "eyJhbGciOiJIUzI1NiJ9.eyJuYW1lIjoiSm9lIENvZGVyIn0.5dlp7GmziL2QS06sZgK4mtaqv0_xX4oFUuTDh1zHK4U"
        "400": 
          description: "error response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "ERROR"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Invalid Login Credentials."
  /api/v1/profile:
    get:
      security:
        - Bearer: []
      tags: 
        - "User Management"
      summary: "Get the current user profile."
      description: "Get the current user profile."
      operationId: "getProfile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        "200": 
          description: "success response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "SUCCESS"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Some message here!"
              additionalData:
                type: "array"
                items:
                  $ref: '#/definitions/TrackrUserDTO'
        "400": 
          description: "error response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "ERROR"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Some message here!"
              additionalData:
                type: "array"
                items:
                  type: "string"
                  example: "some detailed error message"
    put:
      security:
        - Bearer: []    
      tags: 
        - "User Management"
      summary: "Update the current user profile information."
      description: "Update the current user profile information, used primarily in the React front-end."
      operationId: "updateProfile"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/TrackrUserDTO"
      responses:
        "200": 
          description: "success response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "SUCCESS"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Some message here!"
              additionalData:
                type: "array"
                items:
                  $ref: '#/definitions/TrackrUserDTO'
        "400": 
          description: "error response"
          schema:
            type: "object"
            properties:
              status:
                type: "string"
                example: "ERROR"
              date:
                type: "string"
                format: "date"
                example: "2022-06-10"
              message:
                type: "string"
                example: "Some message here!"
              additionalData:
                type: "array"
                items:
                  type: "string"
                  example: "some detailed error message"
...
definitions:
  TrackrUserDTO:
    type: "object"
    properties:
      firstName:
        type: "string"
        example: "John"
      lastName:
        type: "string"
        example: "Doe"
      username:
        type: "string"
        example: "myUsername"
      password:
        type: "string"
        format: "password"
        example: "myCoolPassword"
      email:
        type: "string"
        format: "email"
        example: "myUser@email.com"        
  TokenRetrievalDTO: 
    type: "object"
    properties:
      username:
        type: "string"
        example: "myUsername"
      password:
        type: "string"
        format: "password"
        example: "myCoolPassword"
...